Refactor for extension related files
